refactor(api): admin.tsをリソース別のroutes/admin/配下に分割する
約1900行のadmin.tsをmovies / ceremonies / nominations /
posters / article-linksに分割し、index.tsで従来どおり
単一のadminRoutesに合成する。マウントパスは変更なし。
